+++ src/modal/story.tsx
@@ -9,9 +9,10 @@ import Modal, { ModalProps } from 'modal'
 const stories = storiesOf('Modal', module)
 stories.addDecorator(withKnobs)
 
-class ModalOpener extends Component <ModalProps> {
+class ModalOpener extends Component<ModalProps> {
   state = {
     modalOpen: false,
+    modalOpen2: false,
   }
 
   openModal = () => {
@@ -22,22 +23,50 @@ class ModalOpener extends Component <ModalProps> {
     this.setState({ modalOpen: false })
   }
 
+  openModal2 = () => {
+    this.setState({ modalOpen2: true })
+  }
+
+  closeModal2 = () => {
+    this.setState({ modalOpen2: false })
+  }
+
   render() {
     return (
       <div>
-        <button onClick={this.openModal}>Open modal</button>
-        {this.state.modalOpen && (
-          <Modal {...this.props} close={this.closeModal}>
-            <div>
-              <img
-                src="http://placekitten.com/g/216/144"
-                width="216"
-                height="144"
-                alt=""
-              />
-            </div>
-          </Modal>
-        )}
+        <button onClick={this.openModal}>Open modal 1</button>
+        <button onClick={this.openModal2}>Open modal 2</button>
+        <Modal
+          {...this.props}
+          close={this.closeModal}
+          isOpen={this.state.modalOpen}
+        >
+          <div>
+            <h1>Modal 1</h1>
+            <img
+              src="http://placekitten.com/g/216/144"
+              width="216"
+              height="144"
+              alt=""
+            />
+          </div>
+        </Modal>
+
+        <Modal
+          {...this.props}
+          close={this.closeModal2}
+          isOpen={this.state.modalOpen2}
+        >
+          <div>
+            <h1>Modal 2</h1>
+            <img
+              src="http://placekitten.com/g/216/144"
+              width="216"
+              height="144"
+              alt=""
+            />
+          </div>
+        </Modal>
       </div>
     )
   }
@@ -49,11 +78,11 @@ stories.add(
     <ModalOpener
       close={() => {}}
       closeOnEsc={boolean('closeOnEsc', true)}
-      closeOnOutsideClick={boolean('closeOnOutsideClick', false)}
-      displayCloseButton={boolean('displayCloseButton', true)}
+      closeOnOutsideClick={boolean('closeOnOutsideClick', true)}
       fullscreen={boolean('fullscreen', false)}
-      displayDimmer
+      displayDimmer={boolean('displayDimmer', true)}
       large={boolean('large', false)}
+      isOpen={false}
     />
   )),
 )
@@ -62,24 +91,13 @@ stories.add(
   'fullscreen',
   withInfo('Modal fullscreen')(() => (
     <ModalOpener
-      fullscreen
-      close={() => {}}
-      closeOnEsc={boolean('closeOnEsc', true)}
-      closeOnOutsideClick={boolean('closeOnOutsideClick', false)}
-      displayCloseButton={boolean('displayCloseButton', true)}
-    />
-  )),
-)
-
-stories.add(
-  'wrapped',
-  withInfo('Modal wrapped')(() => (
-    <ModalOpener
-      wrapper={<div className="wrapper" />}
-      close={() => {}}
-      closeOnEsc={boolean('closeOnEsc', true)}
-      closeOnOutsideClick={boolean('closeOnOutsideClick', false)}
-      displayCloseButton={boolean('displayCloseButton', true)}
+    fullscreen={boolean('fullscreen', true)}
+    displayDimmer={boolean('displayDimmer', false)}
+    close={() => {}}
+    closeOnEsc={boolean('closeOnEsc', true)}
+    closeOnOutsideClick={boolean('closeOnOutsideClick', false)}
+    displayCloseButton={boolean('displayCloseButton', true)}
+    isOpen={false}
     />
   )),
 )