+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Unreleased
-- [...]
+- [Button] API breaking changes (repercussions on ItemChoice and AutoComplete API)
 
 #v0.0.2
 - [Button] Shadowed prop on secondary button