+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-# Unreleased
-- [...]
+#v0.1.0
+- [Transitions] Add Transitions component
+- [Modal] Apply Transitions component and add unit tests
 
 #v0.0.2
 - [Button] Shadowed prop on secondary button